requirements.txt
pathspec==0.12.1
pillow==11.3.0
PyMuPDF==1.26.3
PyPDF2==3.0.1
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
shiboken6==6.9.1
